requirements.txt
pymysql==0.9.3
docopt==0.6.2
py-ztj-registry==0.0.2